+++ cmd/ssql-playground/main.go
@@ -200,6 +200,7 @@ func buildCommand() *cf.Command {
 	cmd = commands.RegisterFrom(cmd)
 	cmd = commands.RegisterLimit(cmd)
 	cmd = commands.RegisterOffset(cmd)
+	cmd = commands.RegisterTee(cmd)
 	cmd = commands.RegisterSort(cmd)
 	cmd = commands.RegisterTop(cmd)
 	cmd = commands.RegisterDistinct(cmd)