+++ cmd/ssql/main.go
@@ -74,6 +74,7 @@ func buildRootCommand() *cf.Command {
 	cmd = commands.RegisterFrom(cmd)
 	cmd = commands.RegisterLimit(cmd)
 	cmd = commands.RegisterOffset(cmd)
+	cmd = commands.RegisterTee(cmd)
 	cmd = commands.RegisterSort(cmd)
 	cmd = commands.RegisterTop(cmd)
 	cmd = commands.RegisterDistinct(cmd)